CORE-618 Merged history actions tests into StorageActionTests.
